requirements.txt
apache-airflow
pydantic
requests
psycopg2-binary
pandas
openpyxl
kafka-python==2.0.2
quixstreams